Add script for automating the removal of users due to inactivity
This script will kick users who are in the VA org, or are collaborators
and have not had any activity in the VA org in the last 90 days.

It will also confirm the issue has been open for at least 3 days before
kicking the user, giving them time to reply to the request to preserve.

Signed-off-by: Brett Logan <lindluni@github.com>